+++ plugins/aves_model/lib/src/settings/keys.dart
@@ -41,6 +41,7 @@ class SettingKeys {
   static const mustBackTwiceToExitKey = 'must_back_twice_to_exit';
   static const keepScreenOnKey = 'keep_screen_on';
   static const homePageKey = 'home_page';
+  static const homeCustomCollectionKey = 'home_custom_collection';
   static const enableBottomNavigationBarKey = 'show_bottom_navigation_bar';
   static const confirmCreateVaultKey = 'confirm_create_vault';
   static const confirmDeleteForeverKey = 'confirm_delete_forever';