+++ test_driver/driver_screenshots.dart
@@ -31,6 +31,7 @@ Future<void> configureAndLaunch() async {
     // navigation
     ..keepScreenOn = KeepScreenOn.always
     ..homePage = HomePageSetting.collection
+    ..homeCustomCollection = {}
     ..enableBottomNavigationBar = true
     ..drawerTypeBookmarks = [null, FavouriteFilter.instance]
     // collection